Add python build file
